Move comparison spec code to a separate utility package
